czatex/chat/requirements.txt
Flask
mysql-connector
PyJWT
requests